+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 GAMSPy CHANGELOG
 ================
 
-GAMSPy 1.0.0rc1
----------------
+GAMSPy 0.14.7
+-------------
 - General
   - Include variable infeasibilities in model.computeInfeasibilities().
+  - Remove cone equation type.
   - Fix empty space issue in paths.
 - Documentation
   - Add gamspy probe and gamspy retrieve to the cli reference page.